Rename osm_data_config into config_osm_data
