+++ scripts/config_osm_data.py
